placed Cell and Player classes in seperate folders
