+++ lib/tic_tac_toe/cell.rb
@@ -1,0 +1,8 @@
+module TicTacToe
+    class Cell
+        attr_accessor :value
+        def initialize(value = "")
+            @value = value
+        end
+    end 
+end
