+++ lib/tic_tac_toe/player.rb
@@ -1,0 +1,10 @@
+module TicTacToe
+    class Player
+        attr_reader :name, :color
+        def initialize(value)
+            @name = value.fetch(:name)
+            @color = value.fetch(:color)
+        end
+    end
+    
+end
